pokemon_cv/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pokemon-cv"
version = "0.1.0"
description = "Real-time Pokemon sprite detection and recognition from OBS scenes or webcam feeds"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Pokemon CV Team"}]
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.9",
  "PyYAML>=6.0",
  "torch>=2.2",
  "torchvision>=0.17",
  "faiss-cpu>=1.8.0",
  "scipy>=1.11",
  "tqdm>=4.66",
  "pillow>=10.0",
  "obsws-python>=1.8.0",
  "pygrabber>=0.2",
]

[project.optional-dependencies]
yolo = ["ultralytics>=8.2"]
dev = ["pytest>=8.1", "pytest-cov>=5.0", "ruff>=0.5"]

[project.scripts]
pokemon-cv = "pokemon_cv.app.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]